+++ tests/Dummies/EntryWithNameButNotRegisterableInterface.php
@@ -3,19 +3,17 @@
 declare(strict_types=1);
 
 /**
- * Contains the TestWidgetInterface interface.
+ * Contains the EntryWithNameButNotRegisterableInterface interface.
  *
  * @copyright   Copyright (c) 2023 Attila Fulop
  * @author      Attila Fulop
  * @license     MIT
- * @since       2023-11-30
+ * @since       2023-12-01
  *
  */
 
 namespace Konekt\Extend\Tests\Dummies;
 
-use Konekt\Extend\Contracts\Registerable;
-
-interface TestWidgetInterface extends Registerable
+interface EntryWithNameButNotRegisterableInterface
 {
 }