+++ tests/Dummies/NamedEntry1.php
@@ -3,21 +3,21 @@
 declare(strict_types=1);
 
 /**
- * Contains the TestWidget3 class.
+ * Contains the NamedEntry1 class.
  *
  * @copyright   Copyright (c) 2023 Attila Fulop
  * @author      Attila Fulop
  * @license     MIT
- * @since       2023-11-30
+ * @since       2023-12-01
  *
  */
 
 namespace Konekt\Extend\Tests\Dummies;
 
-class TestWidget3 implements TestWidgetInterface
+class NamedEntry1 implements EntryWithNameButNotRegisterableInterface
 {
     public static function getName(): string
     {
-        return 'Test Widget 3';
+        return 'Named Entry 1';
     }
 }